+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Change Log
 Version 0.21.0 (??????????)
 ---------------------------
 
- * changed pipe symbol from `|` to `>`
+ * changed smooth language: synthetic constructor name is equal to struct name
+ * changed smooth language: changed pipe symbol from `|` to `>`
  * changed smooth language: added anonymous functions
  * redesigned function type literal from A(B,C) to (B,C)->A
  * added parenthesis for enclosing expressions